do not hard depend on symfony templating being activated
